+++ packages/kiarina-lib-cloudflare-d1/src/kiarina/lib/cloudflare/d1/settings.py
@@ -1,8 +1,10 @@
-from pydantic_settings import BaseSettings
+from pydantic_settings import BaseSettings, SettingsConfigDict
 from pydantic_settings_manager import SettingsManager
 
 
 class D1Settings(BaseSettings):
+    model_config = SettingsConfigDict(env_prefix="KIARINA_LIB_CLOUDFLARE_D1_")
+
     database_id: str