+++ packages/kiarina-lib-cloudflare-auth/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Add environment variable prefix `KIARINA_LIB_CLOUDFLARE_AUTH_` to `CloudflareAuthSettings` for better configuration management
+
 ## [1.10.0] - 2025-12-01
 
 ### Changed