+++ packages/kiarina-lib-google-cloud-storage/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Add environment variable prefix `KIARINA_LIB_GOOGLE_CLOUD_STORAGE_` to `GoogleCloudStorageSettings` for better configuration management
+
 ## [1.10.0] - 2025-12-01
 
 ### Changed